Make various CTexture.t2D描画 overloads that perform alignment calculations instead delegate to the new overloads taking Horizontal and Vertical Reference Point arguments.
Note that this does not yet account for those which do so while accounting for texture object scaling ratio (AKA vc拡大縮小倍率.) I need to considre more why they wouldn't want to, since the underlying draw code always accounts for it. It would only seem to "make sense" to opt in and out of that in order to bend the drawing behaviour for effect.